Fix for Donator Flag
There is a bug currently that the Donator Only shows on any server that the setting is edited for. The `/server setting donator` command sets it to 1 in the database no matter what is selected. I believe this is the correct way to fix it (worked in testing on my end, but I'm not a Python expert). I based this on the code used for the SQL update under the `hidden` command.